backend/pyproject.toml
[project]
name = "business_suite"
version = "0.1.0"
description = "A web-based ERP system for managing operations in a visa agency."
requires-python = ">=3.12"
authors = [
    { name = "Stefano Galassi", email = "galaxy73.it@gmail.com" }
]
dependencies = [
    "django>=6.0.0",
    "python-dotenv>=1.2.1",
    "whitenoise>=6.11.0",
    "gunicorn>=23.0.0",
    "setuptools>=80.9.0",
    "huey>=2.5.3",
    "peewee",
    "django-bootstrap-sidebar>=1.2.7",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.3.1",
    "drf-spectacular>=0.27.0",
    "djangorestframework-camel-case>=1.4.0",
    "django-cors-headers>=4.9.0",
    "django-widget-tweaks>=1.5.1",
    "django-nested-admin>=4.1.6",
    "crispy-bootstrap5>=2025.6",
    "passporteye==2.2.2",
    "pdf2image>=1.17.0",
    "django-unicorn>=0.62.0",
    "django-debug-toolbar>=6.1.0",
    "django-waffle>=1.2.1",
    "rapidfuzz>=3.13.0",
    "django-storages>=1.14.6",
    "django-dbbackup>=5.1.0",
    "django-cleanup>=9.0.0",
    "requests>=2.32.5",
    "psycopg2-binary>=2.9.11",
    "dropbox>=12.0.2",
    "pymemcache>=4.0.0",
    "docx-mailmerge>=0.5.0",
    "pyyaml",
    "ua-parser",
    "user-agents",
    "django-user-agents",
    "seaborn",
    "pillow",
    "openai>=1.102.0",
    "openpyxl>=3.1.5",
    "python-docx>=1.2.0",
    "pypdf>=6.5.0",
    "hypothesis>=6.148.9",
    "django-auditlog>=3.1.1",
    "concurrent-log-handler>=0.9.0",
    "google-auth>=2.48.0",
    "google-api-python-client>=2.189.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = [
    "api",
    "core",
    "landing",
    "payments",
    "products",
    "invoices",
    "fixtures",
    "customers",
    "business_suite",
    "customer_applications"
]

[tool.pytest.ini_options]
# Avoid recursing into heavy non-test directories that slow down collection
norecursedirs = [
  ".*",
  ".git",
  ".venv",
  ".hypothesis",
  "node_modules",
  "frontend",
  "frontend/node_modules",
  "frontend/dist",
  "frontend/dist-test",
  "dist",
  "build",
  "tmp",
  "logs",
  "static",
  "public",
  "business_suite.egg-info",
]
# Only consider files that follow the project's test naming convention
python_files = ["test_*.py"]
# Ensure markers are explicitly registered (helps catch typos) and run standard discovery
addopts = "--strict-markers"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
]
